start on fixing warnings so that --no-warnings won't be default

+++ src/simulation/elements/BANG.cpp
@@ -39,7 +39,7 @@ int BANG_update(UPDATE_FUNC_ARGS)
 	}
 	else if(parts[i].tmp == 1)
 	{
-		if ((pmap[y][x]>>8 == i))
+		if ((pmap[y][x]>>8 == (unsigned int)i))
 		{
 			PropertyValue tempValue;
 			tempValue.Integer = 2;
